Move is done.
Everything tested. Bot reached ~1800 ELO against Stockfish.
