+++ src/features/mod.rs
@@ -3,3 +3,4 @@ pub mod killer_moves;
 pub mod opening_book;
 pub mod quiesence;
 pub mod time_management;
+pub mod board_utils;
